Fix server-side rendering.
React was not picking up the markup because it had no data-reactids and it
had no checksum. The markup itself was different too, since the attributes
were being sorted alphabetically (React/om don't do this). These changes make
the "React attempted to reuse markup but..." message on the console
disappear.